refactor: main process architecture — break circular dep, split god objects
SessionManager relocation:
- Move singleton from ipc/session.ts to session/singleton.ts
- Break tools ↔ ipc/session circular dependency
- Remove createRequire lazy-load hacks in tools/index.ts,
  llm/build-prompt-context.ts, llm/tool-dispatch.ts
- Fix 3 layering violations (domain → IPC imports)

SubagentManager splitting (2120 → 1955 lines):
- Extract 7 error classes to agents/errors.ts
- Extract AdmissionController to agents/admission.ts
- Re-exports preserve backward compatibility

registerChatIPC splitting (2113 → 1067 lines, -49.5%):
- Extract 41 module-scope helpers into 8 focused modules under ipc/chat/:
  state.ts, events.ts, snapshot.ts, persist.ts, abort.ts,
  session.ts, stream.ts, title.ts
- registerChatIPC body and nested closures unchanged

All 2910 tests pass.
