+++ electron/src/main/ipc/chat.ts
@@ -7,46 +7,21 @@
  * The chat handler manages an active agent actor per session and
  * forwards StreamEvents as IPC events to the renderer.
  */
-import { ipcMain, webContents as electronWebContents, type WebContents } from 'electron';
+import { ipcMain, type WebContents } from 'electron';
 import type { ReasoningProviderOptions } from '../providers/drivers/types';
-import { createActor, type ActorRefFrom } from 'xstate';
+import { createActor } from 'xstate';
 import { z } from 'zod';
 import { agentMachine, type AgentContext } from '../agents/xstate/agent-machine';
 import { interruptMachine } from '../agents/xstate/interrupt-machine';
-import { streamChat, type StreamEvent } from '../llm/orchestrator';
 import { resolveMainAgentEffort } from '../llm/reasoning-effort';
-import { createMiddlewareStack } from '../llm/middleware';
-import { AgentType, type Agent } from '../../shared/types/agent';
-import type { ModelSelection } from '../../shared/types/provider';
-import { IPC_CHANNELS } from '../../shared/types/ipc';
-import { MessageRole, MessageType } from '../../shared/types/message';
+import { IPC_CHANNELS, type ChatSessionSnapshot } from '../../shared/types/ipc';
+import { MessageType } from '../../shared/types/message';
 import type { Message, Usage } from '../../shared/types/message';
 import { ChainStatus } from '../../shared/types/chain';
-import type { GenerateTitleCallback } from '../session/manager';
-import {
-  flattenSessionMessages,
-  getSessionManager,
-  resolveWindowWorkspace,
-  takeDraftReasoningOverride,
-  takeDraftPermissionOverride,
-} from './session';
-import { workingSetOpenOrFocus } from './session-working-set';
+import { flattenSessionMessages, getSessionManager } from './session';
 import { getBackgroundStore } from '../tools/process/background-store';
 import { getBuiltinToolRegistryForRuntime, getSubagentManager } from '../tools';
-import type {
-  ChatErrorKind,
-  ChatSendResult,
-  ChatSnapshot,
-  ChatSessionSnapshot,
-  ChatSnapshotState,
-  ChatStreamSegmentSnapshot,
-  ChatToolCallSnapshot,
-} from '../../shared/types/ipc';
-import {
-  clearAllChatHistory,
-  getChatHistory,
-  setChatHistory,
-} from './chat-history';
+import { clearAllChatHistory, getChatHistory } from './chat-history';
 import {
   makeAssistantMessage,
   makeThinkingMessage,
@@ -54,19 +29,13 @@ import {
   makeToolResultMessage,
   makeUserMessage,
 } from '../llm/message-factories';
-import { clearDraftCwd } from '../project/workspace';
 import type { ToolExecutionContext } from '../tools/types';
-import {
-  getProjectRuntimeRegistry,
-  type ProjectRuntime,
-} from '../project/runtime';
 import {
   completeSessionActivity,
   publishSessionActivity,
 } from './session-activity';
 import { appendProjectPersonality } from '../project/personality';
 import { appendRootAgentsMd, findRootAgentsMdEntry } from '../project/agents-md';
-import { buildSystemPromptContext } from '../llm/build-prompt-context';
 import {
   acquireProjectMCPManager,
   releaseProjectMCPManager,
@@ -75,8 +44,6 @@ import { getProviderRuntime } from '../providers';
 import type { LanguageModelV4 } from '@ai-sdk/provider';
 import { getProviderAccountingStore } from '../providers/accounting/store';
 import type { ProviderAttemptAccountingContext } from '../providers/accounting/middleware';
-import { importESM } from '../utils/esm-import';
-import { getTierModelSelection } from '../config/loader';
 import {
   chatCancelSchema,
   chatQueueNextSchema,
@@ -87,8 +54,59 @@ import {
 import {
   clearNextRequestStop,
   requestNextRequestStop,
-  shouldStopNextRequest,
 } from './next-request-stop';
+import {
+  activeAgents,
+  agentGenerations,
+  canEmitStreamEvents,
+  draftEnsureByWindow,
+  isCurrentAgent,
+  nextAgentGeneration,
+  sessionsStarting,
+  type ActiveAgent,
+} from './chat/state';
+import {
+  canSend,
+  emitSessionUpdated,
+  sendChatState,
+  sendTurnEvent,
+  webContentsForWindowId,
+} from './chat/events';
+import {
+  appendTextSegment,
+  ensureToolSnapshot,
+  snapshotForAgent,
+  textSegmentIdAtOffset,
+  updateToolSnapshot,
+} from './chat/snapshot';
+import {
+  appendLiveTailMessages,
+  attachUsageToLatestAssistant,
+  checkpointActiveTurn,
+  flushPartialTurnContent,
+  historyFromSession,
+  persistTurnConversation,
+  turnMessagesFromAgent,
+} from './chat/persist';
+import {
+  disposeActiveAgent,
+  forceAbortSession,
+  forceStopSession,
+} from './chat/abort';
+import { ensureActiveSessionSingleFlight } from './chat/session';
+import { classifyErrorKind, createProviderStreamFn } from './chat/stream';
+import { createGenerateTitleCallback } from './chat/title';
+
+export { getActiveMainTurnWindowId, getLiveChatSnapshot } from './chat/snapshot';
+export {
+  activeSessionsForProviderConnection,
+  forceAbortChat,
+  forceAbortMainTurn,
+  stopActiveProviderConnectionTurns,
+} from './chat/abort';
+export type { ForceAbortMainTurnOptions } from './chat/abort';
+export { ensureActiveSession } from './chat/session';
+export { forceAbortSession, forceStopSession, webContentsForWindowId };
 
 /** Upper bound for bgcmd:snapshot lastN (prevents unbounded payload reads). */
 const BG_CMD_SNAPSHOT_MAX_LAST_N = 1000;
@@ -99,1070 +117,6 @@ const bgCommandSnapshotSchema = z.object({
   /** Owning session; when omitted, resolved from the calling window's active session. */
   sessionId: z.string().uuid().optional(),
 });
-
-// ── Active actor tracking ────────────────────────────────────────────────────
-
-type ActiveAgent = {
-  /** Stable execution owner; all persistence and events are addressed by it. */
-  sessionId: string;
-  /** Window that initiated the turn (used only as an event destination). */
-  windowId: string;
-  /** Chain/turn identity used to order snapshot and stream events. */
-  turnId: string;
-  /** Project working directory frozen when this turn began. */
-  cwd: string;
-  /** Epoch timestamp for live elapsed-time displays. */
-  startedAt: number;
-  actor: ActorRefFrom<typeof agentMachine>;
-  interruptActor: ActorRefFrom<typeof interruptMachine>;
-  abortController: AbortController;
-  /**
-   * Full conversation history for the LLM at the start of this turn
-   * (prior turns flattened + current user message).
-   */
-  messages: Message[];
-  /**
-   * Count of messages from prior turns only (before this turn's user message).
-   * Turn-local chain messages = messages.slice(priorMessageCount) + turnMessages.
-   */
-  priorMessageCount: number;
-  /** Messages produced during this turn (tool calls/results + assistant). */
-  turnMessages: Message[];
-  /** Length of context.response already snapshotted into turnMessages as text. */
-  responseCommittedLength: number;
-  /** Length of context.thinking already snapshotted into turnMessages. */
-  thinkingCommittedLength: number;
-  agent: Agent;
-  /** Connection-scoped selection frozen for this turn's chain/storage. */
-  selection: ModelSelection;
-  agentCancelled: boolean;
-  finalized: boolean;
-  /** Monotonic generation for this window; stale agents must not emit IPC. */
-  generation: number;
-  /** Per-turn sequence used to order live events against snapshots. */
-  eventSequence: number;
-  /** Last steady-state metadata sent to renderers; content travels via chunks. */
-  lastChatState: ChatStatePayload | null;
-  /** Current tool cards, kept outside the renderer so a session can rehydrate. */
-  toolCalls: Map<string, ChatToolCallSnapshot>;
-  /** Chronological live timeline (text, thinking, and tools) for rehydration. */
-  streamSegments: ChatStreamSegmentSnapshot[];
-  unsubscribe: () => void;
-  interruptUnsubscribe: () => void;
-  interruptResetTimer: ReturnType<typeof setTimeout> | null;
-  /** Releases turn-scoped resources exactly once when the actor is disposed. */
-  releaseResources: () => void;
-};
-
-type ChatStatePayload = {
-  state: string;
-  error: string | null;
-  interruptState: 'idle' | 'confirmAgent' | 'confirmSubagents';
-  cwd: string | null;
-};
-
-const activeAgents = new Map<string, ActiveAgent>();
-const sessionsStarting = new Set<string>();
-const pendingCheckpoints = new Map<string, { timer: ReturnType<typeof setTimeout>; messages: Message[] }>();
-/**
- * Single-flight draft session create per window. Concurrent first sends from
- * draft mode share one in-flight ensure promise so only one session is created.
- */
-const draftEnsureByWindow = new Map<
-  string,
-  Promise<{
-    ok: true;
-    cwd: string;
-    session: import('../../shared/types/session').Session;
-    runtime: ProjectRuntime;
-  } | { ok: false; result: ChatSendResult }>
->();
-
-/**
- * Per-session generation counter. Incremented on every new chat:send and on
- * forceAbort so stale actor/interrupt subscriptions can drop events even if
- * they fire after the agent was replaced or torn down.
- */
-const agentGenerations = new Map<string, number>();
-
-function nextAgentGeneration(sessionId: string): number {
-  const gen = (agentGenerations.get(sessionId) ?? 0) + 1;
-  agentGenerations.set(sessionId, gen);
-  return gen;
-}
-
-function disposeActiveAgent(sessionId: string, active: ActiveAgent): void {
-  cancelPendingCheckpoint(sessionId);
-  // Only clear the map slot if we still own it (a newer agent may have replaced us).
-  if (activeAgents.get(sessionId) === active) {
-    activeAgents.delete(sessionId);
-  }
-  active.unsubscribe();
-  active.interruptUnsubscribe();
-  if (active.interruptResetTimer) {
-    clearTimeout(active.interruptResetTimer);
-    active.interruptResetTimer = null;
-  }
-  active.abortController.abort();
-  active.actor.stop();
-  active.interruptActor.stop();
-  active.releaseResources();
-}
-
-/**
- * Whether this agent may still stream IPC to the renderer.
- * Drops events from cancelled, finalized, replaced, or generation-stale agents.
- */
-function canEmitStreamEvents(sessionId: string, active: ActiveAgent): boolean {
-  return (
-    !active.agentCancelled &&
-    !active.finalized &&
-    activeAgents.get(sessionId) === active &&
-    agentGenerations.get(sessionId) === active.generation
-  );
-}
-
-/** True when this agent still occupies the window's active slot (may be cancelled). */
-function isCurrentAgent(sessionId: string, active: ActiveAgent): boolean {
-  return (
-    activeAgents.get(sessionId) === active &&
-    agentGenerations.get(sessionId) === active.generation
-  );
-}
-
-function nextEventIdentity(active: ActiveAgent) {
-  active.eventSequence += 1;
-  return {
-    sessionId: active.sessionId,
-    turnId: active.turnId,
-    sequence: active.eventSequence,
-  };
-}
-
-function sendTurnEvent(
-  webContents: WebContents,
-  active: ActiveAgent,
-  channel: string,
-  payload: Record<string, unknown>,
-): void {
-  const identity = nextEventIdentity(active);
-  const recipients = new Map<number, WebContents>();
-  recipients.set(webContents.id, webContents);
-  const allWebContents = electronWebContents.getAllWebContents?.() ?? [];
-  for (const candidate of allWebContents) {
-    if (getSessionManager().getActive(String(candidate.id))?.id === active.sessionId) {
-      recipients.set(candidate.id, candidate);
-    }
-  }
-  for (const recipient of recipients.values()) {
-    if (canSend(recipient)) {
-      recipient.send(channel, { ...identity, ...payload });
-    }
-  }
-}
-
-/**
- * Send state only when renderer-visible metadata changes. Streaming content is
- * intentionally excluded: it is delivered as CHAT_CHUNK and retained in the
- * explicit ChatSnapshot hydration path.
- */
-function sendChatState(
-  webContents: WebContents,
-  active: ActiveAgent,
-  payload: ChatStatePayload,
-): void {
-  const previous = active.lastChatState;
-  if (
-    previous &&
-    previous.state === payload.state &&
-    previous.error === payload.error &&
-    previous.interruptState === payload.interruptState &&
-    previous.cwd === payload.cwd
-  ) {
-    return;
-  }
-
-  active.lastChatState = payload;
-  sendTurnEvent(webContents, active, IPC_CHANNELS.CHAT_STATE, payload);
-}
-
-function appendTextSegment(
-  active: ActiveAgent,
-  kind: 'text' | 'thinking',
-  content: string,
-): string {
-  const last = active.streamSegments.at(-1);
-  if (last?.kind === kind) {
-    active.streamSegments[active.streamSegments.length - 1] = {
-      ...last,
-      content: last.content + content,
-    };
-    return last.id;
-  }
-  const id = crypto.randomUUID();
-  active.streamSegments.push({ kind, id, content });
-  return id;
-}
-
-function textSegmentIdAtOffset(
-  active: ActiveAgent,
-  kind: 'text' | 'thinking',
-  offset: number,
-): string | undefined {
-  let consumed = 0;
-  for (const segment of active.streamSegments) {
-    if (segment.kind !== kind) continue;
-    const end = consumed + segment.content.length;
-    if (end > offset) return segment.id;
-    consumed = end;
-  }
-  return undefined;
-}
-
-function ensureToolSnapshot(
-  active: ActiveAgent,
-  toolCallId: string,
-  toolName: string,
-): ChatToolCallSnapshot {
-  const existing = active.toolCalls.get(toolCallId);
-  if (existing) return existing;
-  const next: ChatToolCallSnapshot = {
-    toolCallId,
-    toolName,
-    status: 'generating',
-    partialArgs: '',
-    args: '',
-    content: null,
-    toolResult: null,
-    startedAt: new Date().toISOString(),
-    finishedAt: null,
-  };
-  active.toolCalls.set(toolCallId, next);
-  active.streamSegments.push({ kind: 'tool', toolCallId });
-  return next;
-}
-
-function updateToolSnapshot(
-  active: ActiveAgent,
-  toolCallId: string,
-  toolName: string,
-  patch: Partial<ChatToolCallSnapshot>,
-): ChatToolCallSnapshot {
-  const existing = ensureToolSnapshot(active, toolCallId, toolName);
-  const next = { ...existing, ...patch, toolCallId };
-  active.toolCalls.set(toolCallId, next);
-  return next;
-}
-
-function snapshotState(active: ActiveAgent): ChatSnapshotState {
-  const state = String(active.actor.getSnapshot().value);
-  if (state === 'error') return 'error';
-  if (state === 'idle') return 'idle';
-  return 'streaming';
-}
-
-function snapshotForAgent(active: ActiveAgent): ChatSnapshot {
-  const context = active.actor.getSnapshot().context as AgentContext;
-  const interruptState = active.interruptActor.getSnapshot().value as
-    | 'idle'
-    | 'confirmAgent'
-    | 'confirmSubagents';
-  return {
-    sessionId: active.sessionId,
-    turnId: active.turnId,
-    sequence: active.eventSequence,
-    state: snapshotState(active),
-    response: context.response ?? '',
-    thinking: context.thinking ?? '',
-    toolCalls: [...active.toolCalls.values()].map((tool) => ({ ...tool })),
-    streamSegments: active.streamSegments.map((segment) => ({ ...segment })),
-    usage: (context.usage as Usage | null) ?? null,
-    error: context.error ?? null,
-    interruptState,
-    cwd: active.cwd,
-    startedAt: active.startedAt,
-    interrupted: active.agentCancelled,
-  };
-}
-
-/**
- * In-flight turn snapshot for a session, or null when idle/finalized.
- *
- * Shared by chat:snapshot and session:open so both hydrate the renderer from
- * one source. Finalization and persistence share one synchronous callback, so
- * a finalized actor is already represented by history when IPC observes it.
- */
-export function getLiveChatSnapshot(sessionId: string): ChatSnapshot | null {
-  const active = activeAgents.get(sessionId);
-  return active && !active.finalized ? snapshotForAgent(active) : null;
-}
-
-/** Originating renderer window for the active main-agent turn. */
-export function getActiveMainTurnWindowId(sessionId: string): string | null {
-  const active = activeAgents.get(sessionId);
-  return active && !active.finalized ? active.windowId : null;
-}
-
-function attachUsageToLatestAssistant(messages: Message[], usage: Usage): boolean {
-  for (let index = messages.length - 1; index >= 0; index--) {
-    const message = messages[index];
-    if (message?.role === MessageRole.ASSISTANT && message.type === MessageType.TEXT) {
-      messages[index] = { ...message, usage };
-      return true;
-    }
-  }
-  return false;
-}
-
-/** Append the current uncommitted live tail to a turn-local message snapshot. */
-function appendLiveTailMessages(
-  messages: Message[],
-  agent: ActiveAgent,
-  context: Pick<AgentContext, 'response' | 'thinking' | 'usage'> | undefined,
-  opts?: { placeholderWhenEmpty?: boolean },
-): void {
-  const partialResponse = context?.response ?? '';
-  const thinking = context?.thinking ?? '';
-  const usage = context?.usage ?? null;
-
-  if (thinking.length > agent.thinkingCommittedLength) {
-    const segment = thinking.slice(agent.thinkingCommittedLength);
-    if (segment.trim()) {
-      messages.push(makeThinkingMessage(
-        segment,
-        textSegmentIdAtOffset(agent, 'thinking', agent.thinkingCommittedLength),
-      ));
-    }
-  }
-
-  const remaining = partialResponse.slice(agent.responseCommittedLength);
-  if (remaining) {
-    messages.push(makeAssistantMessage(
-      remaining,
-      usage,
-      textSegmentIdAtOffset(agent, 'text', agent.responseCommittedLength),
-    ));
-  } else if (opts?.placeholderWhenEmpty && messages.length === 0) {
-    messages.push(makeAssistantMessage(
-      partialResponse,
-      usage,
-      textSegmentIdAtOffset(agent, 'text', agent.responseCommittedLength),
-    ));
-  } else if (usage && !attachUsageToLatestAssistant(messages, usage)) {
-    messages.push({
-      ...makeAssistantMessage('', usage),
-      hidden: true,
-    });
-  }
-}
-
-/**
- * Flush partial stream content into turnMessages (thinking + uncommitted
- * assistant text). Shared by forceAbort, replace-on-send, and error paths.
- */
-function flushPartialTurnContent(agent: ActiveAgent, context: AgentContext | undefined): void {
-  const partialResponse = context?.response ?? '';
-  const thinking = context?.thinking ?? '';
-  appendLiveTailMessages(agent.turnMessages, agent, context);
-  if (thinking.length > agent.thinkingCommittedLength) {
-    agent.thinkingCommittedLength = thinking.length;
-  }
-  if (partialResponse.length > agent.responseCommittedLength) {
-    agent.responseCommittedLength = partialResponse.length;
-  }
-}
-
-/** Resolve WebContents for a window id (forceAbort / SESSION_UPDATED). */
-export function webContentsForWindowId(windowId: string): WebContents | null {
-  try {
-    const id = Number(windowId);
-    if (!Number.isFinite(id)) return null;
-    const wc = electronWebContents.fromId(id);
-    if (!wc || (typeof wc.isDestroyed === 'function' && wc.isDestroyed())) {
-      return null;
-    }
-    return wc;
-  } catch {
-    return null;
-  }
-}
-
-/**
- * Silently abort any in-flight chat for a window (e.g. on session switch).
- * Does not emit CHAT_DONE — the renderer is about to replace its message list.
- *
- * Dispose is synchronous: a deferred microtask left a window where the old
- * subscription could still emit CHAT_CHUNK after session:load swapped UI state
- * (or after a new chat:send started). Flags + generation bump drop any late
- * callbacks that race with stop/unsubscribe.
- *
- * Before discarding, we attempt to persist any partial turn (user message +
- * tool calls + assistant text produced so far) as INTERRUPTED so the user does
- * not lose context when switching sessions mid-stream (P2-9).
- *
- * If the agent already finalized (persist completed), only dispose — never
- * mint a duplicate INTERRUPTED chain.
- */
-export function forceAbortChat(windowId: string): void {
-  const sessionId = getSessionManager().getActive(windowId)?.id;
-  if (sessionId) forceAbortSession(sessionId);
-}
-
-/** Abort exactly one session without affecting work in any other session. */
-export function forceAbortSession(sessionId: string): void {
-  getBackgroundStore().terminateSession(sessionId);
-  try {
-    getSubagentManager().cancelRunning(sessionId);
-  } catch (err) {
-    console.debug(
-      'forceAbortSession subagent cancel failed (non-fatal):',
-      err,
-    );
-  }
-
-  forceAbortMainTurn(sessionId);
-}
-
-export interface ForceAbortMainTurnOptions {
-  /** Notify the originating renderer that its visible turn was interrupted. */
-  emitTerminalEvents?: boolean;
-}
-
-/**
- * Abort and persist only the main-agent turn for one session.
- *
- * Interactive-question cancellation uses this narrower path so declining a
- * main-agent question cannot terminate independent subagents or background
- * commands owned by the same session.
- */
-export function forceAbortMainTurn(
-  sessionId: string,
-  options: ForceAbortMainTurnOptions = {},
-): void {
-  const existing = activeAgents.get(sessionId);
-  if (!existing) return;
-
-  // Already finalized (done/error/cancel) — only dispose; do not re-persist.
-  if (existing.finalized) {
-    existing.agentCancelled = true;
-    nextAgentGeneration(sessionId);
-    disposeActiveAgent(sessionId, existing);
-    return;
-  }
-
-  let context: AgentContext | undefined;
-  try {
-    const snapshot = existing.actor.getSnapshot();
-    context = snapshot?.context as AgentContext | undefined;
-    flushPartialTurnContent(existing, context);
-
-    if (existing.messages.length > 0 || existing.turnMessages.length > 0) {
-      const fullHistory = [...existing.messages, ...existing.turnMessages];
-      if (fullHistory.length > 0) {
-        try {
-          const wc = webContentsForWindowId(existing.windowId);
-          persistTurnConversation(
-            sessionId,
-            fullHistory,
-            turnMessagesFromAgent(existing),
-            ChainStatus.INTERRUPTED,
-            existing.agent,
-            existing.selection,
-            wc ?? undefined,
-          );
-        } catch (err) {
-          console.debug(
-            'Failed to persist partial chat on forceAbort (non-fatal):',
-            err,
-          );
-        }
-      }
-    }
-  } catch (err) {
-    console.debug(
-      'forceAbortMainTurn persistence attempt failed (non-fatal):',
-      err,
-    );
-  }
-
-  existing.agentCancelled = true;
-  existing.finalized = true;
-  completeSessionActivity(
-    sessionId,
-    getSessionManager().getActive(existing.windowId)?.id !== sessionId,
-  );
-
-  if (options.emitTerminalEvents) {
-    const ownerWebContents = webContentsForWindowId(existing.windowId);
-    if (ownerWebContents) {
-      const response = context?.response ?? '';
-      const usage = context?.usage ?? null;
-      try {
-        sendTurnEvent(ownerWebContents, existing, IPC_CHANNELS.CHAT_DONE, {
-          type: 'done',
-          response,
-          interrupted: true,
-          usage,
-        });
-      } catch (err) {
-        console.debug('Failed to emit CHAT_DONE on main-turn abort (non-fatal):', err);
-      }
-      try {
-        sendChatState(ownerWebContents, existing, {
-          state: 'idle',
-          error: null,
-          interruptState: 'idle',
-          cwd: existing.cwd,
-        });
-      } catch (err) {
-        console.debug('Failed to emit CHAT_STATE on main-turn abort (non-fatal):', err);
-      }
-    }
-  }
-
-  nextAgentGeneration(sessionId);
-  disposeActiveAgent(sessionId, existing);
-}
-
-/**
- * Immediately stop one session for the global Activity surface.
- *
- * Unlike Esc cancellation, this does not require confirmation clicks. It keeps
- * all writes and terminal events on the stopped session's originating window.
- */
-export function forceStopSession(sessionId: string): boolean {
-  getBackgroundStore().terminateSession(sessionId);
-  const existing = activeAgents.get(sessionId);
-  const cancelledSubagents = getSubagentManager().cancelRunning(sessionId);
-  if (!existing) {
-    if (cancelledSubagents.length > 0) {
-      completeSessionActivity(sessionId, true);
-    }
-    return cancelledSubagents.length > 0;
-  }
-
-  // Already finalized (done/error/cancel) — only dispose residual work.
-  if (existing.finalized) {
-    existing.agentCancelled = true;
-    nextAgentGeneration(sessionId);
-    disposeActiveAgent(sessionId, existing);
-    return true;
-  }
-
-  const ownerWebContents =
-    webContentsForWindowId(existing.windowId) ?? null;
-  existing.agentCancelled = true;
-  existing.finalized = true;
-  const context = existing.actor.getSnapshot().context as AgentContext;
-  existing.actor.send({ type: 'CANCEL' });
-  flushPartialTurnContent(existing, context);
-  const fullHistory = [...existing.messages, ...existing.turnMessages];
-  if (fullHistory.length > 0) {
-    persistTurnConversation(
-      sessionId,
-      fullHistory,
-      turnMessagesFromAgent(existing),
-      ChainStatus.INTERRUPTED,
-      existing.agent,
-      existing.selection,
-      ownerWebContents ?? undefined,
-    );
-  }
-  completeSessionActivity(
-    sessionId,
-    getSessionManager().getActive(existing.windowId)?.id !== sessionId,
-  );
-
-  if (ownerWebContents) {
-    sendTurnEvent(ownerWebContents, existing, IPC_CHANNELS.CHAT_DONE, {
-      type: 'done',
-      response: context.response ?? '',
-      interrupted: true,
-      usage: context.usage ?? null,
-    });
-    sendChatState(ownerWebContents, existing, {
-      state: 'idle',
-      error: null,
-      interruptState: 'idle',
-      cwd: existing.cwd,
-    });
-  }
-
-  nextAgentGeneration(sessionId);
-  disposeActiveAgent(sessionId, existing);
-  return true;
-}
-
-/** Active session IDs whose frozen turn uses the given provider connection. */
-export function activeSessionsForProviderConnection(connectionId: string): readonly string[] {
-  return [...activeAgents.values()]
-    .filter((active) => !active.finalized && active.selection.connectionId === connectionId)
-    .map((active) => active.sessionId);
-}
-
-/**
- * Destructive disconnect helper. Stops only turns already attributed to this
- * connection; other connections and frozen completed turns remain untouched.
- */
-export function stopActiveProviderConnectionTurns(connectionId: string): readonly string[] {
-  const sessionIds = activeSessionsForProviderConnection(connectionId);
-  for (const sessionId of sessionIds) forceStopSession(sessionId);
-  return sessionIds;
-}
-
-function canSend(webContents: WebContents): boolean {
-  return typeof webContents.isDestroyed !== 'function' || !webContents.isDestroyed();
-}
-
-type EnsureActiveSessionResult =
-  | {
-      ok: true;
-      cwd: string;
-      session: import('../../shared/types/session').Session;
-      runtime: ProjectRuntime;
-    }
-  | { ok: false; result: ChatSendResult };
-
-/**
- * Ensure there is a session ready before streaming/persisting.
- * Draft mode leaves no active session until the first chat:send — create
- * lazily here and notify the renderer so the sidebar gains a list entry.
- *
- * When `requestedSessionId` is set, resolves that session by id without
- * changing the window's active selection (mid-flight / background sends).
- *
- * Requires a valid workspace (draft → session → sticky default). Never uses
- * process.cwd() as the product default. If unbound, does not create a session.
- *
- * @returns ok + session cwd, or a structured failure for the send gate
- */
-export function ensureActiveSession(
-  webContents: WebContents,
-  preferredModel?: ModelSelection | null,
-  requestedSessionId?: string,
-  draftGeneration?: number,
-): EnsureActiveSessionResult {
-  const windowId = String(webContents.id);
-  const manager = getSessionManager();
-  // Resolve by id without switchTo — do not steal window selection mid-flight.
-  let active = requestedSessionId
-    ? manager.getSession(requestedSessionId)
-    : manager.getActive(windowId);
-  if (requestedSessionId && !active) {
-    return {
-      ok: false,
-      result: {
-        status: 'error',
-        error: 'The requested session no longer exists.',
-        kind: 'session_not_found',
-      },
-    };
-  }
-  const workspace = resolveWindowWorkspace(windowId);
-
-  const boundCwd = active?.cwd?.trim() || workspace.cwd;
-
-  if (boundCwd == null || boundCwd === '') {
-    return {
-      ok: false,
-      result: {
-        status: 'error',
-        error:
-          'No project folder selected. Choose a folder before sending a message.',
-        kind: 'unbound_workspace',
-      },
-    };
-  }
-
-  // Resolve once at turn start. The returned snapshot is independent from
-  // whatever project another window selects while this turn is running.
-  const runtime = getProjectRuntimeRegistry().get(boundCwd);
-
-  const selection = preferredModel ?? active?.selection ?? runtime.config.default_model;
-  if (selection == null) {
-    return {
-      ok: false,
-      result: {
-        status: 'error',
-        error: 'A provider connection and model are required before sending a message.',
-        kind: 'provider_required',
-      },
-    };
-  }
-
-  // Draft path: re-check in case a concurrent first-send just created a session.
-  if (!active) {
-    active = manager.getActive(windowId);
-  }
-
-  if (active) {
-    const selectedNow = manager.getActive(windowId)?.id === active.id;
-    if (preferredModel && (
-      active.selection?.connectionId !== preferredModel.connectionId
-      || active.selection?.modelId !== preferredModel.modelId
-    )) {
-      if (selectedNow) {
-        manager.changeModel(active.id, preferredModel, preferredModel.modelId);
-        active = manager.getSession(active.id) ?? { ...active, selection: preferredModel };
-      } else {
-        // Turn-local override only — do not steal selection to persist.
-        active = { ...active, selection: preferredModel };
-      }
-    }
-    return { ok: true, cwd: boundCwd, session: active, runtime };
-  }
-
-  const created = manager.create(
-    selection,
-    { cwd: boundCwd },
-    windowId,
-    selection.modelId,
-  );
-  // A reasoning effort chosen in draft mode rides into the new session so the
-  // very first turn (which reads the returned session) already honors it.
-  const draftOverride = takeDraftReasoningOverride(windowId);
-  if (draftOverride !== undefined) {
-    manager.setReasoningEffortOverride(created.id, draftOverride);
-  }
-  const draftPermission = takeDraftPermissionOverride(windowId);
-  if (draftPermission !== undefined) {
-    manager.setPermissionMode(created.id, draftPermission);
-  }
-  const session =
-    manager.getSession(created.id) ??
-    { ...created, reasoningEffortOverride: draftOverride ?? created.reasoningEffortOverride };
-  // Draft was promoted into the new session.
-  clearDraftCwd(windowId);
-  workingSetOpenOrFocus(session.id, windowId);
-  if (canSend(webContents)) {
-    webContents.send(IPC_CHANNELS.SESSION_CREATED, { session, draftGeneration });
-  }
-  return { ok: true, cwd: boundCwd, session, runtime };
-}
-
-/**
- * Window-level single-flight for draft first-send. Concurrent chat:send without
- * sessionId share one ensure promise so only one session is created.
- */
-function ensureActiveSessionSingleFlight(
-  webContents: WebContents,
-  preferredModel?: ModelSelection | null,
-  requestedSessionId?: string,
-  draftGeneration?: number,
-): EnsureActiveSessionResult | Promise<EnsureActiveSessionResult> {
-  const windowId = String(webContents.id);
-  const manager = getSessionManager();
-  // Existing session or explicit id: no draft create race.
-  if (requestedSessionId || manager.getActive(windowId)) {
-    return ensureActiveSession(
-      webContents,
-      preferredModel,
-      requestedSessionId,
-      draftGeneration,
-    );
-  }
-
-  const inflight = draftEnsureByWindow.get(windowId);
-  if (inflight) return inflight;
-
-  let resolveFlight!: (value: EnsureActiveSessionResult) => void;
-  let rejectFlight!: (reason: unknown) => void;
-  const flight = new Promise<EnsureActiveSessionResult>((resolve, reject) => {
-    resolveFlight = resolve;
-    rejectFlight = reject;
-  });
-  draftEnsureByWindow.set(windowId, flight);
-
-  try {
-    const result = ensureActiveSession(
-      webContents,
-      preferredModel,
-      requestedSessionId,
-      draftGeneration,
-    );
-    resolveFlight(result);
-  } catch (error) {
-    rejectFlight(error);
-    draftEnsureByWindow.delete(windowId);
-    throw error;
-  } finally {
-    queueMicrotask(() => {
-      if (draftEnsureByWindow.get(windowId) === flight) {
-        draftEnsureByWindow.delete(windowId);
-      }
-    });
-  }
-  return flight;
-}
-
-function classifyErrorKind(title: string | null | undefined, detail: string): ChatErrorKind {
-  const haystack = `${title ?? ''} ${detail}`.toLowerCase();
-  if (haystack.includes('rate limit') || haystack.includes('429') || haystack.includes('usage limit')) {
-    return 'rate-limit';
-  }
-  if (
-    haystack.includes('auth') ||
-    haystack.includes('401') ||
-    haystack.includes('403') ||
-    haystack.includes('api key')
-  ) {
-    return 'auth';
-  }
-  if (
-    haystack.includes('timeout') ||
-    haystack.includes('timed out') ||
-    haystack.includes('network') ||
-    haystack.includes('connection')
-  ) {
-    return 'stream';
-  }
-  return 'generic';
-}
-
-/**
- * Build turn-local messages for multi-chain storage:
- * current user message (+ any pre-turn messages after priorMessageCount) +
- * tool/assistant messages produced during the turn.
- */
-function turnMessagesFromAgent(agent: ActiveAgent): Message[] {
-  const turnBase = agent.messages.slice(agent.priorMessageCount);
-  return [...turnBase, ...agent.turnMessages];
-}
-
-/** Materialize the current live tail without mutating committed turn state. */
-function checkpointMessagesFromAgent(agent: ActiveAgent, context: AgentContext): Message[] {
-  const checkpoint = turnMessagesFromAgent(agent);
-  appendLiveTailMessages(checkpoint, agent, context);
-  return checkpoint;
-}
-
-const CHECKPOINT_DEBOUNCE_MS = 300;
-
-/** Persist one bounded main-turn checkpoint, debounced per session. */
-function checkpointActiveTurn(agent: ActiveAgent, context: AgentContext): void {
-  const sessionId = agent.sessionId;
-  const messages = checkpointMessagesFromAgent(agent, context);
-  const existing = pendingCheckpoints.get(sessionId);
-  if (existing) {
-    existing.messages = messages;
-    return;
-  }
-  const timer = setTimeout(() => {
-    const entry = pendingCheckpoints.get(sessionId);
-    pendingCheckpoints.delete(sessionId);
-    const active = activeAgents.get(sessionId);
-    if (!active || active.finalized) return;
-    try {
-      getSessionManager().updateActiveChainMessages(entry?.messages ?? messages, sessionId);
-    } catch (err) {
-      console.debug('Failed to checkpoint active chat chain (non-fatal):', err);
-    }
-  }, CHECKPOINT_DEBOUNCE_MS);
-  pendingCheckpoints.set(sessionId, { timer, messages });
-}
-
-/** Cancel any pending debounced checkpoint for a session. */
-function cancelPendingCheckpoint(sessionId: string): void {
-  const pending = pendingCheckpoints.get(sessionId);
-  if (pending) {
-    clearTimeout(pending.timer);
-    pendingCheckpoints.delete(sessionId);
-  }
-}
-
-/**
- * Persist flat LLM history + turn-local multi-chain write.
- *
- * - Window chatHistory keeps the full flattened conversation for the next send.
- * - Session storage writes only `turnMessages` onto the ACTIVE chain (or
- *   creates one via persistTurn when startChain was skipped).
- */
-function persistTurnConversation(
-  sessionId: string,
-  fullHistory: Message[],
-  turnMessages: Message[],
-  status: ChainStatus,
-  agent: Agent,
-  selection?: ModelSelection | null,
-  webContents?: WebContents,
-): void {
-  setChatHistory(sessionId, fullHistory);
-  try {
-    const sessionManager = getSessionManager();
-    const updated = sessionManager.persistTurn({
-      messages: turnMessages,
-      status,
-      selection,
-      modelLabel: selection?.modelId ?? null,
-      agentName: agent.name,
-      agentType: agent.type,
-      agentTier: agent.tier,
-    }, sessionId);
-    if (updated && webContents && canSend(webContents)) {
-      webContents.send(IPC_CHANNELS.SESSION_UPDATED, { session: updated });
-    }
-  } catch (err) {
-    console.debug('Failed to persist chat chain (non-fatal):', err);
-  }
-}
-
-/** Flatten all session chains — never only the active/last chain. */
-function historyFromSession(sessionId: string): Message[] {
-  try {
-    const session = getSessionManager().getSession(sessionId);
-    if (!session) return [];
-    return flattenSessionMessages(session);
-  } catch {
-    return [];
-  }
-}
-
-/** Notify renderer of live session (multi-chain) state after startChain. */
-function emitSessionUpdated(webContents: WebContents, sessionId: string): void {
-  try {
-    const session = getSessionManager().getSession(sessionId);
-    if (session && canSend(webContents)) {
-      webContents.send(IPC_CHANNELS.SESSION_UPDATED, { session });
-    }
-  } catch {
-    // non-fatal
-  }
-}
-
-// ── Stream function (wraps the orchestrator) ─────────────────────────────────
-
-/**
- * Bind a turn's already-resolved adapter to the orchestrator. The typed
- * selection, project runtime, message history, and model instance are all
- * frozen before the actor starts, so a later settings change cannot redirect
- * credentials, tools, or a retry to another connection.
- */
-function createProviderStreamFn(input: {
-  readonly messages: Message[];
-  readonly runtime: ProjectRuntime;
-  readonly sessionId: string;
-  readonly windowId: string;
-  readonly modelInstance: LanguageModelV4;
-  readonly accounting: ProviderAttemptAccountingContext;
-  readonly registry: ReturnType<typeof getBuiltinToolRegistryForRuntime>;
-  readonly mcpManager: ReturnType<typeof acquireProjectMCPManager>;
-  readonly providerOptions?: ReasoningProviderOptions;
-}) {
-  return async function* ({
-    agent,
-    systemPrompt,
-    abortSignal,
-  }: {
-    message: string;
-    agent: Agent;
-    systemPrompt: string;
-    abortSignal: AbortSignal;
-  }): AsyncGenerator<StreamEvent> {
-    const context = await buildSystemPromptContext({
-      cwd: input.runtime.projectDir,
-      config: input.runtime.config,
-      sessionId: input.sessionId,
-      agentScopeId: 'main',
-    });
-    yield* streamChat({
-      messages: input.messages,
-      agent,
-      systemPrompt,
-      context,
-      config: input.runtime.config,
-      registry: input.registry,
-      mcpManager: input.mcpManager,
-      sessionId: input.sessionId,
-      windowId: input.windowId,
-      projectRuntime: input.runtime,
-      agentScopeId: 'main',
-      abortSignal,
-      shouldStopEarly: () => shouldStopNextRequest(input.sessionId),
-      modelInstance: input.modelInstance,
-      accounting: input.accounting,
-      providerOptions: input.providerOptions,
-    });
-  };
-}
-
-// ── Auto-naming callback factory ─────────────────────────────────────────────
-
-const SESSION_NAMER_AGENT_NAME = 'session-namer';
-
-/**
- * Creates a GenerateTitleCallback that uses the bundled internal session-namer
- * agent to produce a short title from the first user/assistant exchange.
- *
- * Non-fatal on failure — returns null so the session keeps its default name.
- */
-function createGenerateTitleCallback(input: {
-  runtime: ProjectRuntime;
-  messages: readonly Message[];
-  fallbackSelection: ModelSelection;
-  accounting: Omit<ProviderAttemptAccountingContext, 'snapshot'>;
-}): GenerateTitleCallback {
-  return async () => {
-    const userMessage = input.messages.find(
-      (message) => message.role === MessageRole.USER && message.type === MessageType.TEXT,
-    );
-    const assistantMessage = input.messages.find(
-      (message) => message.role === MessageRole.ASSISTANT && message.type === MessageType.TEXT,
-    );
-    if (!userMessage || !assistantMessage) {
-      console.warn(
-        '[auto-name] Completed exchange has no user/assistant text; keeping the default session name.',
-      );
-      return null;
-    }
-
-    try {
-      const titleAgent = input.runtime.agents.get(SESSION_NAMER_AGENT_NAME);
-      if (!titleAgent || titleAgent.type !== AgentType.INTERNAL) {
-        console.warn(
-          `[auto-name] Internal agent "${SESSION_NAMER_AGENT_NAME}" is unavailable; ` +
-          'keeping the default session name.',
-        );
-        return null;
-      }
-      const titleSelection =
-        getTierModelSelection(input.runtime.config, titleAgent.tier) ??
-        input.fallbackSelection;
-      const execution = await getProviderRuntime().resolveExecution(titleSelection);
-      const { generateText, wrapLanguageModel } = await importESM<typeof import('ai')>('ai');
-      const model = wrapLanguageModel({
-        model: execution.modelInstance,
-        middleware: createMiddlewareStack({
-          retry: { maxRetries: input.runtime.config.llm_stream_retries },
-          accounting: { ...input.accounting, snapshot: execution.snapshot },
-        }),
-      });
-      const result = await generateText({
-        model,
-        instructions: titleAgent.system_prompt,
-        abortSignal: AbortSignal.timeout(
-          Math.max(1, input.runtime.config.llm_stream_idle_timeout * 1000),
-        ),
-        messages: [
-          {
-            role: 'user',
-            content:
-              `User: ${userMessage.content.slice(0, 500)}\n\n` +
-              `Assistant: ${assistantMessage.content.slice(0, 500)}`,
-          },
-        ],
-        // Orchid's accounting-aware retry middleware owns every retry attempt.
-        maxRetries: 0,
-      });
-      return result.text.trim().split('\n')[0]?.trim() || null;
-    } catch (error) {
-      console.warn(
-        '[auto-name] Title generation failed; keeping the default session name:',
-        error,
-      );
-      return null;
-    }
-  };
-}
 
 // ── IPC registration ─────────────────────────────────────────────────────────
