+++ electron/src/main/tools/subagent/delegate.ts
@@ -20,7 +20,7 @@ import {
 import type { SubagentManager, SubagentRecord } from '../../agents/manager';
 import { SubagentQueueFullError, SubagentState } from '../../agents/manager';
 import { getTierModelSelection } from '../../config/loader';
-import { getSessionManager } from '../../ipc/session';
+import { getSessionManager } from '../../session/singleton';
 
 /**
  * Result returned by all subagent tool handlers.
